// src/generated/models/Dnx_evaluationcriteriasModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dnx_evaluationcriteriasdnx_operator = {
  0: 'Equals',
  1: 'GreaterThan',
  2: 'LessThan',
  3: 'Contains',
  4: 'IsTrue',
  5: 'IsFalse'
} as const;
export type Dnx_evaluationcriteriasdnx_operator = keyof typeof Dnx_evaluationcriteriasdnx_operator;
export const Dnx_evaluationcriteriasdnx_outcome_if_fail = {
  0: 'NotMet',
  1: 'NotSuitable',
  2: 'Fail'
} as const;
export type Dnx_evaluationcriteriasdnx_outcome_if_fail = keyof typeof Dnx_evaluationcriteriasdnx_outcome_if_fail;
export const Dnx_evaluationcriteriasdnx_outcome_if_pass = {
  0: 'Met',
  1: 'Suitable',
  2: 'Pass'
} as const;
export type Dnx_evaluationcriteriasdnx_outcome_if_pass = keyof typeof Dnx_evaluationcriteriasdnx_outcome_if_pass;
export const Dnx_evaluationcriteriasdnx_scoring_type = {
  0: 'Weighted',
  1: 'Grouped',
  2: 'Boolean'
} as const;
export type Dnx_evaluationcriteriasdnx_scoring_type = keyof typeof Dnx_evaluationcriteriasdnx_scoring_type;
export const Dnx_evaluationcriteriasdnx_source_type = {
  0: 'QuestionValue',
  1: 'SubsectionOutcome',
  2: 'SectionOutcome',
  3: 'DocumentExtract'
} as const;
export type Dnx_evaluationcriteriasdnx_source_type = keyof typeof Dnx_evaluationcriteriasdnx_source_type;
export const Dnx_evaluationcriteriasstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dnx_evaluationcriteriasstatecode = keyof typeof Dnx_evaluationcriteriasstatecode;
export const Dnx_evaluationcriteriasstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dnx_evaluationcriteriasstatuscode = keyof typeof Dnx_evaluationcriteriasstatuscode;

export interface Dnx_evaluationcriteriasBase {
  "dnx_Assessment_Level@odata.bind"?: string;
  dnx_criteria_name: string;
  dnx_evaluationcriteriaid: string;
  dnx_operator?: Dnx_evaluationcriteriasdnx_operator;
  dnx_outcome_if_fail?: Dnx_evaluationcriteriasdnx_outcome_if_fail;
  dnx_outcome_if_pass?: Dnx_evaluationcriteriasdnx_outcome_if_pass;
  dnx_pass_threshold?: number;
  dnx_scoring_type?: Dnx_evaluationcriteriasdnx_scoring_type;
  "dnx_Source_Assessment_Level@odata.bind"?: string;
  dnx_source_type?: Dnx_evaluationcriteriasdnx_source_type;
  dnx_target_value?: string;
  dnx_weight?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dnx_evaluationcriteriasstatecode;
  statuscode?: Dnx_evaluationcriteriasstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dnx_evaluationcriterias extends Dnx_evaluationcriteriasBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dnx_assessment_levelname?: string;
  dnx_operatorname?: string;
  dnx_outcome_if_failname?: string;
  dnx_outcome_if_passname?: string;
  dnx_scoring_typename?: string;
  dnx_source_assessment_levelname?: string;
  dnx_source_typename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dnx_assessment_level?: object;
  _dnx_assessment_level_value?: string;
  dnx_source_assessment_level?: object;
  _dnx_source_assessment_level_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}